Add ALEGoToImplementation
In languages, like C and C++, there is a declaration for a function or
method, and there is a definition.  The former, which is usually placed
in a header file, is reached with the LSP Goto Definition Request [sic].
However, the latter, which is usually separate, is implemented with a
different request, the Goto Implementation Request.  Let's add an
ALEGoToImplementation command that uses this request to jump to a C or
C++ definition.

In many other languages, like Rust, the two commands are functionally
equivalent.